activity_bar.pl: Swap privmsg and regular message colours
After using the plugin for a couple days, I decided that orange is too
insistent a colour to describe messages I probably don't need to read
immediately.

Swap the blue and orange colours so that now blue means regular message
and orange means private message. This feels better in terms of message
priority.